Add --watch-extra options to check non page dirs
Awestruct ignores any change made on _data or _partials. Because these are
extensions, directories are customizable.
I have added a --watch-extra option
accepting a comma separated list of directories to watch for.
When a file is changed, the site is regenerated (like what is done for site.yml)

A typical usage would be `awestruct -d --watch-extra _data,_partials`