Add FSSM support to development mode
Replaced wild loop with File monitor (hooks into OS file monitors e.g. inotify)

Will add a monitor to all subdirectories and files that does not match:
  /^(_site|_tmp|\.git|\.gitignore|\.sass-cache|\.|\.\.).*/

A CREATE or UPDATE will cause run on the Generation command to be called.

Note:

No CREATE events will be caught in the root directory it self, since the directory is not watched, but the individual files. Could not find a way to exclude a directory via glob in root monitor, e.g. '**/*^_site'