Use :type in the deploy configuration.
This allows for other kinds of deployment types, such as openshift to be
plugged in. This is not backwards compatible with existing github pages
deployments. Any sites that are currently deploying to github pages will
need to update the deploy section of the profile so that it includes a
type of `:github_pages` and an optional branch name, the default being
`gh-pages`.