ContextHelper#fully_qualify_urls fix for issue #112.
Allow anchor tags to omit an href attribute. This allows for anchor
targets such as <a name="foo"/>. Also move development dependencies from
the gemspec file to the Gemfile in a development group. This may help
with issue #123 and running on Windows, but will probably require a bit
more tweaking in the Gemfile mayhap.