Spawn regeneration thread in Dev mode
Spawn a new thread when the Guard listener detect a change on a monitored file.
The new thread will perform the normal site regeneration, but the Guard listener
is now free to receive new events.

If a new change event on the same path as currently running is detected, the running
thread(if any) will be killed, and a new thread is spawned that starts the
regeneration again. If a different path is changed, the started regeneration is
allowed to finish before the new path is regenerated.

This allow the user to make multiple rapid changes to the same template without having
to wait for the re-generation of all the dependent pages to finish before the
current working template is regenerated.