Add individual tracking of content and variables
Changes to Front Matter variables are only read once and updates are
never reflected to the Page object as they can be changed by the pipline
which only run during startup. Tracking dependencies between Page objects
for both content and variables as one is wastefull in this scenario as
a lot of Pages will depend on each others variables, but very few on their
content. Regenerating a Page because a dependency has changed Front Matter
has no effect on the output (the old variables are used).

Instead track content and variable dependencies individually, but only
trigger regeneration of dependent Pages if content has changed.
The content and Variables are hashed using SHA256 which allow us to detect
a real change, and not just the a file has changed modify time.

When a change event occure, the following steps are taken:

* Locate the Page for the given Change event Path
* Trigger regeneration of that Page
* If that Page's Content has changed
    * Find all dependent Pages and trigger regeneration
* If that Page's variables have changed
    * Warn that a restart is required for the change to take effect

The corresponding Page to a changed path will always be generated first,
before its depending pages. The depending pages are sorted by last modified
time. This allow the regeneration process to start with 'the most interesting'
pages first to give the user a quick feedback on change, while updating
the complete dependent tree in the 'background'.

> Page variable dependencies are still tracked, but not currently used.
> A possible usecase could be to display a Page's dependencies as debug info.