Move :html, :js and :css compression from using cmd to ruby gems
The minfiy extension rely on calling yuicompressor and htmlcompressor on
the command line with no control if they are installed or not. Change Minify
to use the htmlcompressor and yui-compressor ruby gems to control the
dependencies.

pngcrush is untouched and still rely on calling the cmd.

Includes :html, :css and :js compression spec tests.