Adding 'selected_tag' as an argument to the paginator
This 'selected_tag' variable becomes available in the page context if
the page is generated by the tagger. This enables the generated page
to display the current tag in the page body.

For example, if /news/tags/awestruct/index.html is generated, the page
can display somewhere in its body that theselected tag is: "awestruct"