Tests pass on haml now
Finally found the correct combination for jruby and mri
Also fixed a bug in a test I think. Front matter is delineated by "---"
not "--". That's what the front_matter_handler is telling me. With only
"--" we have a page, which is not what the test was.

Removed a file I shouldn't have committed anyway