Adding new extension hooks
There are three new extension hooks in place now:

* before_extensions
+
All extensions added via this method, following the same extension rules, will be
run before any other extensions loaded with `extension`.
* after_extensions
+
All extensions added via this method, following the same extension rules, will be
run after any other extensions loaded with `extension`.
* after_generation
+
All extensions added via this method, following the same extension rules, will be
run after the page generation is completed. This will only run on the
first generation, not on reloads.

Extensions will now be notified if they are being reloaded. To be
notified of this an extension must have an `on_reload(site)` method,
which will be called if there is a change. An extension is then free to
re-run an initialization logic, clean-up the site variable, or create
new state needed for a re-run of the `execute(site)` method.