Switch between in process or in_thread generation
There's a new configuration option (default shown below) in site.yml:

  generation:
    :in_threads: <%= Parallel.processor_count * 10 %>

This controls the parallelization for page output generation.
`generation` must be a hash, and accepts any Parallel options available
with the version of Parallel being used, see
https://github.com/grosser/parallel for more information.

The two main options for this are `:in_threads: <number of threads to
use>` and `:in_processes: <number of processes to use>`. The
`Parallel.processor_count` is a method which returns the number of
processors available. Both of these two options expect an integer to be
returned.

WARNING: Make sure to use the preceding ':' in the key to make in a
symbol, otherwise things won't work correctly.