Fixing up the broken tagger extension
I broke Tagger with the change to enable frontmatter changes. I this is
a slighty better approach, but may not cover all cases. Now we're
updating the page variables if they keys are the same and the classes
are the same. If what we need to update is an Array or a Hash we're
checking the classes of each value to see if they match. If they do
match then we'll update it. This should catch things like tags which is
in the front matter, but we change it through an extension.

I also remove an extra loop inside the Tagger extension should increase
perf a little bit.