Improving the tracing output
There were issues of the ending of the json being clipped, not sure why,
adding padding at end seems to fix it. May only be an issue on Webrick
though.

If an object isn't showing up correctly in the trace you can implement
`to_h` to have it show up as an object.

If an object has no keys, it's an empty object, show `{}` in the trace
output.