Forcefully close link tags in atom template
Otherwise the generated feed won't actually be valid XML! Eep!  Haml defaults
to 'html5' which *opens* self-closing tags in the Haml markup whether you want
it to or not

One other potential workaround, which I like less, is to change the format to
:xhtml, e.g.

```
Haml::Options.defaults[:format] = :xhtml
```

This is unfortunately a global configuration setting and I've yet to find a way
to change the format on a per-template basis.

By including the "raw" <link/> tags in the template, we bypass that level of
the Haml engine