Fixes #511 (performance logging)
I had to redo the logging infrastructure of awestruct as it didn't allow for what we wanted to do here easily. We're now using the logging gem, which is very similar to log4j and log4j. This gives us a much more flexible logging solution.

Also I STRONGLY recommend AGAINST using debug (-D, --debug) for large sites. This produces A LOT OF DATA which you probably will never need or want